ORM class 2 participation.

+++ ORM/01-Activities/12-Stu_RESTful-Routes/Unsolved/routes/api/userRoutes.js
@@ -6,30 +6,60 @@ const User = require('../../models/User');
 
 // GET a user
 router.get('/:id', async (req, res) => {
-  const userData = await User.findByPk(req.params.id).catch((err) =>
-    res.json(err)
-  );
-  res.json(userData);
+  try {
+    const userData = await User.findByPk(req.params.id);
+
+    if (!userData) {
+      return res.status(404).json({ message: 'No user found with this ID!' });
+    }
+
+    res.json(userData);
+  } catch (error) {
+    res.status(500).json({ error });
+  };
+
 });
 
 // UPDATE a user
 router.put('/:id', async (req, res) => {
-  const userData = await User.update(req.body, {
-    where: {
-      id: req.params.id,
-    },
-  }).catch((err) => res.json(err));
-  res.json(userData);
+  try {
+    const doesUserExist = await User.findByPk(req.params.id);
+    if (!doesUserExist) {
+      return res.status(404).json({ message: 'No user found with this ID!' });
+    }
+
+    const userData = await User.update(req.body, {
+      where: {
+        id: req.params.id,
+      },
+    });
+
+    res.json(userData);
+
+  } catch (error) {
+    res.status(500).json({ error });
+  }
+
 });
 
 // DELETE a user
 router.delete('/:id', async (req, res) => {
-  const userData = await User.destroy({
-    where: {
-      id: req.params.id,
-    },
-  }).catch((err) => res.json(err));
-  res.json(userData);
+  try {
+    const doesUserExist = await User.findByPk(req.params.id);
+    if (!doesUserExist) {
+      return res.status(404).json({ message: 'No user found with this ID!' });
+    }
+
+    const userData = await User.destroy({
+      where: {
+        id: req.params.id,
+      },
+    });
+    res.json(userData);
+
+  } catch (error) {
+    res.status(500).json({ error });
+  }
 });
 
 module.exports = router;
